Revert "Feature/fix v1 controllers --> release"
