+++ controllers/campo.go
@@ -5,20 +5,21 @@ import (
 	"errors"
 	"strconv"
 	"strings"
-
-	"github.com/udistrital/oikos_api/models"
+	"time"
 
 	"github.com/astaxie/beego"
 	"github.com/astaxie/beego/logs"
+
+	"github.com/udistrital/oikos_api/models"
 )
 
-// CampoV2Controller operations for Campo
-type CampoV2Controller struct {
+// CampoController oprations for Campo
+type CampoController struct {
 	beego.Controller
 }
 
 // URLMapping ...
-func (c *CampoV2Controller) URLMapping() {
+func (c *CampoController) URLMapping() {
 	c.Mapping("Post", c.Post)
 	c.Mapping("GetOne", c.GetOne)
 	c.Mapping("GetAll", c.GetAll)
@@ -29,14 +30,26 @@ func (c *CampoV2Controller) URLMapping() {
 // Post ...
 // @Title Post
 // @Description create Campo
-// @Param	body		body 	models.CampoV2	true		"body for Campo content"
-// @Success 201 {object} models.CampoV2
-// @Failure 400 the request contains incorrect syntax
+// @Param	body		body 	models.Campo	true		"body for Campo content"
+// @Success 201 {object} models.Campo
+// @Failure 403 body is empty
 // @router / [post]
-func (c *CampoV2Controller) Post() {
-	var v models.CampoV2
+func (c *CampoController) Post() {
+	var v models.Campo
 	if err := json.Unmarshal(c.Ctx.Input.RequestBody, &v); err == nil {
-		if _, err := models.AddCampo(&v); err == nil {
+		//-------------- Temporal: Cambio por transición ------- //
+
+		temp := models.CampoV2{
+			Id:                v.Id,
+			Nombre:            v.Nombre,
+			Descripcion:       v.Descripcion,
+			CodigoAbreviacion: "C_" + v.Nombre,
+			Activo:            true,
+			FechaCreacion:     time.Now(),
+			FechaModificacion: time.Now(),
+		}
+		if _, err := models.AddCampo(&temp); err == nil {
+			//if _, err := models.AddCampo(&v); err == nil {
 			c.Ctx.Output.SetStatus(201)
 			c.Data["json"] = v
 		} else {
@@ -58,10 +71,10 @@ func (c *CampoV2Controller) Post() {
 // @Title Get One
 // @Description get Campo by id
 // @Param	id		path 	int	true		"The key for staticblock"
-// @Success 200 {object} models.CampoV2
-// @Failure 404 not found resource
+// @Success 200 {object} models.Campo
+// @Failure 403 :id is empty
 // @router /:id [get]
-func (c *CampoV2Controller) GetOne() {
+func (c *CampoController) GetOne() {
 	idStr := c.Ctx.Input.Param(":id")
 	id, _ := strconv.Atoi(idStr)
 	v, err := models.GetCampoById(id)
@@ -69,9 +82,22 @@ func (c *CampoV2Controller) GetOne() {
 		logs.Error(err)
 		//c.Data["development"] = map[string]interface{}{"Code": "000", "Body": err.Error(), "Type": "error"}
 		c.Data["system"] = err
-		c.Abort("404")
+		c.Abort("400")
 	} else {
-		c.Data["json"] = v
+
+		temp := models.Campo{
+			Id:                v.Id,
+			Nombre:            v.Nombre,
+			Descripcion:       v.Descripcion,
+			CodigoAbreviacion: v.CodigoAbreviacion,
+			Activo:            v.Activo,
+			FechaCreacion:     v.FechaCreacion,
+			FechaModificacion: v.FechaModificacion,
+		}
+
+		c.Data["json"] = temp
+		//-------------- Temporal: Cambio por transición ------- //
+		//c.Data["json"] = v
 	}
 	c.ServeJSON()
 }
@@ -85,10 +111,10 @@ func (c *CampoV2Controller) GetOne() {
 // @Param	order	query	string	false	"Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc ..."
 // @Param	limit	query	int	false	"Limit the size of result set. Must be an integer"
 // @Param	offset	query	int	false	"Start position of result set. Must be an integer"
-// @Success 200 {object} []models.CampoV2
-// @Failure 404 not found resource
+// @Success 200 {object} []models.Campo
+// @Failure 403
 // @router / [get]
-func (c *CampoV2Controller) GetAll() {
+func (c *CampoController) GetAll() {
 	var fields []string
 	var sortby []string
 	var order []string
@@ -139,8 +165,29 @@ func (c *CampoV2Controller) GetAll() {
 	} else {
 		if l == nil {
 			l = append(l, map[string]interface{}{})
+			c.Data["json"] = l
+		} else {
+			//-------------- Temporal: Cambio por transición ------- //
+			var temp []models.Campo
+			for _, i := range l {
+				field, _ := i.(models.CampoV2)
+
+				x := models.Campo{
+					Id:                field.Id,
+					Nombre:            field.Nombre,
+					Descripcion:       field.Descripcion,
+					CodigoAbreviacion: field.CodigoAbreviacion,
+					Activo:            field.Activo,
+					FechaCreacion:     field.FechaCreacion,
+					FechaModificacion: field.FechaModificacion,
+				}
+
+				temp = append(temp, x)
+			}
+			c.Data["json"] = temp
 		}
-		c.Data["json"] = l
+
+		//c.Data["json"] = l
 	}
 	c.ServeJSON()
 }
@@ -149,16 +196,28 @@ func (c *CampoV2Controller) GetAll() {
 // @Title Put
 // @Description update the Campo
 // @Param	id		path 	int	true		"The id you want to update"
-// @Param	body		body 	models.CampoV2	true		"body for Campo content"
-// @Success 200 {object} models.CampoV2
-// @Failure 400 the request contains incorrect syntax
+// @Param	body		body 	models.Campo	true		"body for Campo content"
+// @Success 200 {object} models.Campo
+// @Failure 403 :id is not int
 // @router /:id [put]
-func (c *CampoV2Controller) Put() {
+func (c *CampoController) Put() {
 	idStr := c.Ctx.Input.Param(":id")
 	id, _ := strconv.Atoi(idStr)
-	v := models.CampoV2{Id: id}
+	infoDep, _ := models.GetCampoById(id)
+	v := models.Campo{Id: id}
+
 	if err := json.Unmarshal(c.Ctx.Input.RequestBody, &v); err == nil {
-		if err := models.UpdateCampoById(&v); err == nil {
+		v2 := models.CampoV2{
+			Id:                id,
+			Nombre:            v.Nombre,
+			Descripcion:       v.Descripcion,
+			CodigoAbreviacion: infoDep.CodigoAbreviacion,
+			Activo:            infoDep.Activo,
+			FechaCreacion:     infoDep.FechaCreacion,
+			FechaModificacion: time.Now(),
+		}
+
+		if err := models.UpdateCampoById(&v2); err == nil {
 			c.Data["json"] = v
 		} else {
 			logs.Error(err)
@@ -180,9 +239,9 @@ func (c *CampoV2Controller) Put() {
 // @Description delete the Campo
 // @Param	id		path 	int	true		"The id you want to delete"
 // @Success 200 {object} models.Deleted
-// @Failure 404 not found resource
+// @Failure 403 id is empty
 // @router /:id [delete]
-func (c *CampoV2Controller) Delete() {
+func (c *CampoController) Delete() {
 	idStr := c.Ctx.Input.Param(":id")
 	id, _ := strconv.Atoi(idStr)
 	if err := models.DeleteCampo(id); err == nil {
