+++ controllers/dependencia_padre.go
@@ -5,6 +5,7 @@ import (
 	"errors"
 	"strconv"
 	"strings"
+	"time"
 
 	"github.com/astaxie/beego"
 	"github.com/astaxie/beego/logs"
@@ -12,32 +13,53 @@ import (
 	"github.com/udistrital/oikos_api/models"
 )
 
-// DependenciaPadreV2Controller operations for DependenciaPadre
-type DependenciaPadreV2Controller struct {
+// DependenciaPadreController oprations for DependenciaPadre
+type DependenciaPadreController struct {
 	beego.Controller
 }
 
 // URLMapping ...
-func (c *DependenciaPadreV2Controller) URLMapping() {
+func (c *DependenciaPadreController) URLMapping() {
 	c.Mapping("Post", c.Post)
 	c.Mapping("GetOne", c.GetOne)
 	c.Mapping("GetAll", c.GetAll)
 	c.Mapping("Put", c.Put)
 	c.Mapping("Delete", c.Delete)
-
+	c.Mapping("FacultadesConProyectos", c.FacultadesConProyectos)
+	c.Mapping("ArbolDependencias", c.ArbolDependencias)
 }
 
 // Post ...
 // @Title Post
 // @Description create DependenciaPadre
-// @Param	body		body 	models.DependenciaPadreV2	true		"body for DependenciaPadre content"
-// @Success 201 {object} models.DependenciaPadreV2
+// @Param	body		body 	models.DependenciaPadre	true		"body for DependenciaPadre content"
+// @Success 201 {object} models.DependenciaPadre
 // @Failure 400 the request contains incorrect syntax
 // @router / [post]
-func (c *DependenciaPadreV2Controller) Post() {
-	var v models.DependenciaPadreV2
+func (c *DependenciaPadreController) Post() {
+	var v models.DependenciaPadre
 	if err := json.Unmarshal(c.Ctx.Input.RequestBody, &v); err == nil {
-		if _, err := models.AddDependenciaPadre(&v); err == nil {
+		//-------------- Temporal: Cambio por transición ------- //
+
+		dp := &models.DependenciaV2{
+			Id: v.Padre.Id,
+		}
+
+		dh := &models.DependenciaV2{
+			Id: v.Hija.Id,
+		}
+
+		temp := models.DependenciaPadreV2{
+			Id:                v.Id,
+			PadreId:           dp,
+			HijaId:            dh,
+			Activo:            true,
+			FechaCreacion:     time.Now(),
+			FechaModificacion: time.Now(),
+		}
+		//-------------- Temporal: Cambio por transición ------- //
+		if _, err := models.AddDependenciaPadre(&temp); err == nil {
+			//if _, err := models.AddDependenciaPadre(&v); err == nil {
 			c.Ctx.Output.SetStatus(201)
 			c.Data["json"] = v
 		} else {
@@ -59,10 +81,10 @@ func (c *DependenciaPadreV2Controller) Post() {
 // @Title Get One
 // @Description get DependenciaPadre by id
 // @Param	id		path 	int	true		"The key for staticblock"
-// @Success 200 {object} models.DependenciaPadreV2
+// @Success 200 {object} models.DependenciaPadre
 // @Failure 404 not found resource
 // @router /:id [get]
-func (c *DependenciaPadreV2Controller) GetOne() {
+func (c *DependenciaPadreController) GetOne() {
 	idStr := c.Ctx.Input.Param(":id")
 	id, _ := strconv.Atoi(idStr)
 	v, err := models.GetDependenciaPadreById(id)
@@ -72,7 +94,32 @@ func (c *DependenciaPadreV2Controller) GetOne() {
 		c.Data["system"] = err
 		c.Abort("404")
 	} else {
-		c.Data["json"] = v
+		//-------------- Temporal: Cambio por transición ------- //
+		dp := &models.Dependencia{
+			Id:                  v.PadreId.Id,
+			Nombre:              v.PadreId.Nombre,
+			TelefonoDependencia: v.PadreId.TelefonoDependencia,
+			CorreoElectronico:   v.PadreId.CorreoElectronico,
+		}
+
+		dh := &models.Dependencia{
+			Id:                  v.HijaId.Id,
+			Nombre:              v.HijaId.Nombre,
+			TelefonoDependencia: v.HijaId.TelefonoDependencia,
+			CorreoElectronico:   v.HijaId.CorreoElectronico,
+		}
+
+		temp := models.DependenciaPadre{
+			Id:                v.Id,
+			Padre:             dp,
+			Hija:              dh,
+			Activo:            v.Activo,
+			FechaCreacion:     v.FechaCreacion,
+			FechaModificacion: v.FechaModificacion,
+		}
+
+		c.Data["json"] = temp
+		//		c.Data["json"] = v
 	}
 	c.ServeJSON()
 }
@@ -86,10 +133,10 @@ func (c *DependenciaPadreV2Controller) GetOne() {
 // @Param	order	query	string	false	"Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc ..."
 // @Param	limit	query	int	false	"Limit the size of result set. Must be an integer"
 // @Param	offset	query	int	false	"Start position of result set. Must be an integer"
-// @Success 200 {object} []models.DependenciaPadreV2
+// @Success 200 {object} []models.DependenciaPadre
 // @Failure 404 not found resource
 // @router / [get]
-func (c *DependenciaPadreV2Controller) GetAll() {
+func (c *DependenciaPadreController) GetAll() {
 	var fields []string
 	var sortby []string
 	var order []string
@@ -140,8 +187,44 @@ func (c *DependenciaPadreV2Controller) GetAll() {
 	} else {
 		if l == nil {
 			l = append(l, map[string]interface{}{})
+			c.Data["json"] = l
+		} else {
+			//-------------- Temporal: Cambio por transición ------- //
+			var temp []models.DependenciaPadre
+			for _, i := range l {
+				field, _ := i.(models.DependenciaPadreV2)
+
+				dp := &models.Dependencia{
+					Id:                  field.PadreId.Id,
+					Nombre:              field.PadreId.Nombre,
+					TelefonoDependencia: field.PadreId.TelefonoDependencia,
+					CorreoElectronico:   field.PadreId.CorreoElectronico,
+				}
+
+				dh := &models.Dependencia{
+					Id:                  field.HijaId.Id,
+					Nombre:              field.HijaId.Nombre,
+					TelefonoDependencia: field.HijaId.TelefonoDependencia,
+					CorreoElectronico:   field.HijaId.CorreoElectronico,
+				}
+
+				x := models.DependenciaPadre{
+					Id:                field.Id,
+					Padre:             dp,
+					Hija:              dh,
+					Activo:            field.Activo,
+					FechaCreacion:     field.FechaCreacion,
+					FechaModificacion: field.FechaModificacion,
+				}
+
+				temp = append(temp, x)
+			}
+
+			c.Data["json"] = temp
+
 		}
-		c.Data["json"] = l
+
+		//c.Data["json"] = l
 	}
 	c.ServeJSON()
 }
@@ -150,16 +233,32 @@ func (c *DependenciaPadreV2Controller) GetAll() {
 // @Title Put
 // @Description update the DependenciaPadre
 // @Param	id		path 	int	true		"The id you want to update"
-// @Param	body		body 	models.DependenciaPadreV2	true		"body for DependenciaPadre content"
-// @Success 200 {object} models.DependenciaPadreV2
+// @Param	body		body 	models.DependenciaPadre	true		"body for DependenciaPadre content"
+// @Success 200 {object} models.DependenciaPadre
 // @Failure 400 the request contains incorrect syntax
 // @router /:id [put]
-func (c *DependenciaPadreV2Controller) Put() {
+func (c *DependenciaPadreController) Put() {
 	idStr := c.Ctx.Input.Param(":id")
 	id, _ := strconv.Atoi(idStr)
-	v := models.DependenciaPadreV2{Id: id}
+	v := models.DependenciaPadre{Id: id}
+	//-------------- Temporal: Cambio por transición ------- //
 	if err := json.Unmarshal(c.Ctx.Input.RequestBody, &v); err == nil {
-		if err := models.UpdateDependenciaPadreById(&v); err == nil {
+		dp := &models.DependenciaV2{
+			Id: v.Padre.Id,
+		}
+		dh := &models.DependenciaV2{
+			Id: v.Hija.Id,
+		}
+		v2 := models.DependenciaPadreV2{
+			Id:                id,
+			PadreId:           dp,
+			HijaId:            dh,
+			Activo:            v.Activo,
+			FechaCreacion:     v.FechaCreacion,
+			FechaModificacion: time.Now(),
+		}
+
+		if err := models.UpdateDependenciaPadreById(&v2); err == nil {
 			c.Data["json"] = v
 		} else {
 			logs.Error(err)
@@ -183,7 +282,7 @@ func (c *DependenciaPadreV2Controller) Put() {
 // @Success 200 {object} models.Deleted
 // @Failure 404 not found resource
 // @router /:id [delete]
-func (c *DependenciaPadreV2Controller) Delete() {
+func (c *DependenciaPadreController) Delete() {
 	idStr := c.Ctx.Input.Param(":id")
 	id, _ := strconv.Atoi(idStr)
 	if err := models.DeleteDependenciaPadre(id); err == nil {
@@ -203,7 +302,7 @@ func (c *DependenciaPadreV2Controller) Delete() {
 // @Success 200 {object} []models.Tree
 // @Failure 403
 // @router /FacultadesConProyectos [get]
-func (c *DependenciaPadreV2Controller) FacultadesConProyectosV2() {
+func (c *DependenciaPadreController) FacultadesConProyectos() {
 	//Construcción Json menus
 	l := models.Facultades()
 	c.Data["json"] = l
@@ -217,7 +316,7 @@ func (c *DependenciaPadreV2Controller) FacultadesConProyectosV2() {
 // @Success 200 {object} []models.TreeDependencia
 // @Failure 403
 // @router /ArbolDependencias [get]
-func (c *DependenciaPadreV2Controller) ArbolDependenciasV2() {
+func (c *DependenciaPadreController) ArbolDependencias() {
 	//Construcción Json menus
 	l := models.ConstruirDependenciasPadre()
 	c.Data["json"] = l
@@ -232,7 +331,7 @@ func (c *DependenciaPadreV2Controller) ArbolDependenciasV2() {
 // @Success 201 {object} models.DependenciaPadreV2
 // @Failure 400 the request contains incorrect syntax
 // @router /tr_dependencia_padre [post]
-func (c *DependenciaPadreV2Controller) TRDependenciaPadreV2() {
+func (c *DependenciaPadreController) TRDependenciaPadre() {
 	var v models.DependenciaPadreV2
 	if err := json.Unmarshal(c.Ctx.Input.RequestBody, &v); err == nil {
