+++ controllers/dependencia_tipo_dependencia.go
@@ -5,20 +5,21 @@ import (
 	"errors"
 	"strconv"
 	"strings"
-
-	"github.com/udistrital/oikos_api/models"
+	"time"
 
 	"github.com/astaxie/beego"
 	"github.com/astaxie/beego/logs"
+
+	"github.com/udistrital/oikos_api/models"
 )
 
-// DependenciaTipoDependenciaV2Controller operations for DependenciaTipoDependencia
-type DependenciaTipoDependenciaV2Controller struct {
+// DependenciaTipoDependenciaController oprations for DependenciaTipoDependencia
+type DependenciaTipoDependenciaController struct {
 	beego.Controller
 }
 
 // URLMapping ...
-func (c *DependenciaTipoDependenciaV2Controller) URLMapping() {
+func (c *DependenciaTipoDependenciaController) URLMapping() {
 	c.Mapping("Post", c.Post)
 	c.Mapping("GetOne", c.GetOne)
 	c.Mapping("GetAll", c.GetAll)
@@ -29,14 +30,34 @@ func (c *DependenciaTipoDependenciaV2Controller) URLMapping() {
 // Post ...
 // @Title Post
 // @Description create DependenciaTipoDependencia
-// @Param	body		body 	models.DependenciaTipoDependenciaV2	true		"body for DependenciaTipoDependencia content"
-// @Success 201 {object} models.DependenciaTipoDependenciaV2
+// @Param	body		body 	models.DependenciaTipoDependencia	true		"body for DependenciaTipoDependencia content"
+// @Success 201 {object} models.DependenciaTipoDependencia
 // @Failure 400 the request contains incorrect syntax
 // @router / [post]
-func (c *DependenciaTipoDependenciaV2Controller) Post() {
-	var v models.DependenciaTipoDependenciaV2
+func (c *DependenciaTipoDependenciaController) Post() {
+	var v models.DependenciaTipoDependencia
 	if err := json.Unmarshal(c.Ctx.Input.RequestBody, &v); err == nil {
-		if _, err := models.AddDependenciaTipoDependencia(&v); err == nil {
+		//-------------- Temporal: Cambio por transición ------- //
+
+		td := &models.TipoDependenciaV2{
+			Id: v.TipoDependenciaId.Id,
+		}
+
+		d := &models.DependenciaV2{
+			Id: v.DependenciaId.Id,
+		}
+
+		temp := models.DependenciaTipoDependenciaV2{
+			Id:                v.Id,
+			TipoDependenciaId: td,
+			DependenciaId:     d,
+			Activo:            true,
+			FechaCreacion:     time.Now(),
+			FechaModificacion: time.Now(),
+		}
+		//-------------- Temporal: Cambio por transición ------- //
+		if _, err := models.AddDependenciaTipoDependencia(&temp); err == nil {
+			//if _, err := models.AddDependenciaTipoDependencia(&v); err == nil {
 			c.Ctx.Output.SetStatus(201)
 			c.Data["json"] = v
 		} else {
@@ -58,10 +79,10 @@ func (c *DependenciaTipoDependenciaV2Controller) Post() {
 // @Title Get One
 // @Description get DependenciaTipoDependencia by id
 // @Param	id		path 	int	true		"The key for staticblock"
-// @Success 200 {object} models.DependenciaTipoDependenciaV2
+// @Success 200 {object} models.DependenciaTipoDependencia
 // @Failure 404 not found resource
 // @router /:id [get]
-func (c *DependenciaTipoDependenciaV2Controller) GetOne() {
+func (c *DependenciaTipoDependenciaController) GetOne() {
 	idStr := c.Ctx.Input.Param(":id")
 	id, _ := strconv.Atoi(idStr)
 	v, err := models.GetDependenciaTipoDependenciaById(id)
@@ -71,7 +92,36 @@ func (c *DependenciaTipoDependenciaV2Controller) GetOne() {
 		c.Data["system"] = err
 		c.Abort("404")
 	} else {
-		c.Data["json"] = v
+		//-------------- Temporal: Cambio por transición ------- //
+		td := &models.TipoDependencia{
+			Id:                v.TipoDependenciaId.Id,
+			Nombre:            v.TipoDependenciaId.Nombre,
+			Descripcion:       v.TipoDependenciaId.Descripcion,
+			CodigoAbreviacion: v.TipoDependenciaId.CodigoAbreviacion,
+			Activo:            v.TipoDependenciaId.Activo,
+			FechaCreacion:     v.TipoDependenciaId.FechaCreacion,
+			FechaModificacion: v.TipoDependenciaId.FechaModificacion,
+		}
+
+		d := &models.Dependencia{
+			Id:                  v.DependenciaId.Id,
+			Nombre:              v.DependenciaId.Nombre,
+			TelefonoDependencia: v.DependenciaId.TelefonoDependencia,
+			CorreoElectronico:   v.DependenciaId.CorreoElectronico,
+		}
+
+		temp := models.DependenciaTipoDependencia{
+			Id:                v.Id,
+			TipoDependenciaId: td,
+			DependenciaId:     d,
+			Activo:            true,
+			FechaCreacion:     time.Now(),
+			FechaModificacion: time.Now(),
+		}
+
+		c.Data["json"] = temp
+
+		//c.Data["json"] = v
 	}
 	c.ServeJSON()
 }
@@ -85,10 +135,10 @@ func (c *DependenciaTipoDependenciaV2Controller) GetOne() {
 // @Param	order	query	string	false	"Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc ..."
 // @Param	limit	query	int	false	"Limit the size of result set. Must be an integer"
 // @Param	offset	query	int	false	"Start position of result set. Must be an integer"
-// @Success 200 {object} []models.DependenciaTipoDependenciaV2
+// @Success 200 {object} []models.DependenciaTipoDependencia
 // @Failure 404 not found resource
 // @router / [get]
-func (c *DependenciaTipoDependenciaV2Controller) GetAll() {
+func (c *DependenciaTipoDependenciaController) GetAll() {
 	var fields []string
 	var sortby []string
 	var order []string
@@ -139,8 +189,45 @@ func (c *DependenciaTipoDependenciaV2Controller) GetAll() {
 	} else {
 		if l == nil {
 			l = append(l, map[string]interface{}{})
+			c.Data["json"] = l
+		} else {
+			//-------------- Temporal: Cambio por transición ------- //
+			var temp []models.DependenciaTipoDependencia
+			for _, i := range l {
+				field, _ := i.(models.DependenciaTipoDependenciaV2)
+
+				td := &models.TipoDependencia{
+					Id:                field.TipoDependenciaId.Id,
+					Nombre:            field.TipoDependenciaId.Nombre,
+					Descripcion:       field.TipoDependenciaId.Descripcion,
+					CodigoAbreviacion: field.TipoDependenciaId.CodigoAbreviacion,
+					Activo:            field.TipoDependenciaId.Activo,
+					FechaCreacion:     field.TipoDependenciaId.FechaCreacion,
+					FechaModificacion: field.TipoDependenciaId.FechaModificacion,
+				}
+
+				d := &models.Dependencia{
+					Id:                  field.DependenciaId.Id,
+					Nombre:              field.DependenciaId.Nombre,
+					TelefonoDependencia: field.DependenciaId.TelefonoDependencia,
+					CorreoElectronico:   field.DependenciaId.CorreoElectronico,
+				}
+
+				x := models.DependenciaTipoDependencia{
+					Id:                field.Id,
+					TipoDependenciaId: td,
+					DependenciaId:     d,
+					Activo:            true,
+					FechaCreacion:     time.Now(),
+					FechaModificacion: time.Now(),
+				}
+
+				temp = append(temp, x)
+			}
+			c.Data["json"] = temp
 		}
-		c.Data["json"] = l
+
+		//c.Data["json"] = l
 	}
 	c.ServeJSON()
 }
@@ -153,10 +240,18 @@ func (c *DependenciaTipoDependenciaV2Controller) GetAll() {
 // @Success 200 {object} models.DependenciaTipoDependenciaV2
 // @Failure 400 the request contains incorrect syntax
 // @router /:id [put]
-func (c *DependenciaTipoDependenciaV2Controller) Put() {
+func (c *DependenciaTipoDependenciaController) Put() {
 	idStr := c.Ctx.Input.Param(":id")
 	id, _ := strconv.Atoi(idStr)
-	v := models.DependenciaTipoDependenciaV2{Id: id}
+	//-------------- Temporal: Cambio por transición ------- //
+	infoDep, _ := models.GetDependenciaTipoDependenciaById(id)
+	v := models.DependenciaTipoDependenciaV2{
+		Id:                id,
+		Activo:            infoDep.Activo,
+		FechaCreacion:     infoDep.FechaCreacion,
+		FechaModificacion: time.Now(),
+	}
+	//v := models.DependenciaTipoDependencia{Id: id}
 	if err := json.Unmarshal(c.Ctx.Input.RequestBody, &v); err == nil {
 		if err := models.UpdateDependenciaTipoDependenciaById(&v); err == nil {
 			c.Data["json"] = v
@@ -182,7 +277,7 @@ func (c *DependenciaTipoDependenciaV2Controller) Put() {
 // @Success 200 {object} models.Deleted
 // @Failure 404 not found resource
 // @router /:id [delete]
-func (c *DependenciaTipoDependenciaV2Controller) Delete() {
+func (c *DependenciaTipoDependenciaController) Delete() {
 	idStr := c.Ctx.Input.Param(":id")
 	id, _ := strconv.Atoi(idStr)
 	if err := models.DeleteDependenciaTipoDependencia(id); err == nil {
