+++ controllers/espacio_fisico.go
@@ -5,6 +5,7 @@ import (
 	"errors"
 	"strconv"
 	"strings"
+	"time"
 
 	"github.com/astaxie/beego"
 	"github.com/astaxie/beego/logs"
@@ -12,31 +13,59 @@ import (
 	"github.com/udistrital/oikos_api/models"
 )
 
-// EspacioFisicoV2Controller operations for EspacioFisico
-type EspacioFisicoV2Controller struct {
+// EspacioFisicoController oprations for EspacioFisico
+type EspacioFisicoController struct {
 	beego.Controller
 }
 
 // URLMapping ...
-func (c *EspacioFisicoV2Controller) URLMapping() {
+func (c *EspacioFisicoController) URLMapping() {
 	c.Mapping("Post", c.Post)
 	c.Mapping("GetOne", c.GetOne)
 	c.Mapping("GetAll", c.GetAll)
 	c.Mapping("Put", c.Put)
 	c.Mapping("Delete", c.Delete)
+	c.Mapping("EspaciosHuerfanos", c.EspaciosHuerfanos)
 }
 
 // Post ...
 // @Title Post
 // @Description create EspacioFisico
-// @Param	body		body 	models.EspacioFisicoV2	true		"body for EspacioFisico content"
-// @Success 201 {object} models.EspacioFisicoV2
+// @Param	body		body 	models.EspacioFisico	true		"body for EspacioFisico content"
+// @Success 201 {object} models.EspacioFisico
 // @Failure 400 the request contains incorrect syntax
 // @router / [post]
-func (c *EspacioFisicoV2Controller) Post() {
-	var v models.EspacioFisicoV2
+func (c *EspacioFisicoController) Post() {
+	var v models.EspacioFisico
 	if err := json.Unmarshal(c.Ctx.Input.RequestBody, &v); err == nil {
-		if _, err := models.AddEspacioFisico(&v); err == nil {
+		//-------------- Temporal: Cambio por transición ------- //
+		var act bool
+		if v.Estado == "Activo" {
+			act = true
+		} else if v.Estado == "Inactivo" {
+			act = false
+		} else {
+			act = true
+		}
+
+		te := &models.TipoEspacioFisicoV2{
+			Id: v.TipoEspacio.Id,
+		}
+
+		temp := models.EspacioFisicoV2{
+			Id:                  v.Id,
+			Nombre:              v.Nombre,
+			Descripcion:         "Descripción - " + v.Nombre,
+			CodigoAbreviacion:   v.Codigo,
+			Activo:              act,
+			FechaCreacion:       time.Now(),
+			FechaModificacion:   time.Now(),
+			TipoEspacioFisicoId: te,
+		}
+
+		if _, err := models.AddEspacioFisico(&temp); err == nil {
+
+			//if _, err := models.AddEspacioFisico(&v); err == nil {
 			c.Ctx.Output.SetStatus(201)
 			c.Data["json"] = v
 		} else {
@@ -58,10 +87,10 @@ func (c *EspacioFisicoV2Controller) Post() {
 // @Title Get One
 // @Description get EspacioFisico by id
 // @Param	id		path 	int	true		"The key for staticblock"
-// @Success 200 {object} models.EspacioFisicoV2
+// @Success 200 {object} models.EspacioFisico
 // @Failure 404 not found resource
 // @router /:id [get]
-func (c *EspacioFisicoV2Controller) GetOne() {
+func (c *EspacioFisicoController) GetOne() {
 	idStr := c.Ctx.Input.Param(":id")
 	id, _ := strconv.Atoi(idStr)
 	v, err := models.GetEspacioFisicoById(id)
@@ -71,7 +100,39 @@ func (c *EspacioFisicoV2Controller) GetOne() {
 		c.Data["system"] = err
 		c.Abort("404")
 	} else {
-		c.Data["json"] = v
+		//-------------- Temporal: Cambio por transición ------- //
+		var act string
+		if v.Activo == true {
+			act = "Activo"
+		} else {
+			act = "Inactivo"
+		}
+
+		te := &models.TipoEspacioFisico{
+			Id:                v.TipoEspacioFisicoId.Id,
+			Nombre:            v.TipoEspacioFisicoId.Nombre,
+			Descripcion:       v.TipoEspacioFisicoId.Descripcion,
+			CodigoAbreviacion: v.TipoEspacioFisicoId.CodigoAbreviacion,
+			Activo:            v.TipoEspacioFisicoId.Activo,
+			FechaCreacion:     v.TipoEspacioFisicoId.FechaCreacion,
+			FechaModificacion: v.TipoEspacioFisicoId.FechaModificacion,
+		}
+
+		temp := models.EspacioFisico{
+			Id:                v.Id,
+			Nombre:            v.Nombre,
+			Codigo:            v.CodigoAbreviacion,
+			Estado:            act,
+			Descripcion:       v.Descripcion,
+			FechaCreacion:     v.FechaCreacion,
+			FechaModificacion: v.FechaModificacion,
+			TipoEspacio:       te,
+			//espacio_fisicoTipoespacio_fisico: field.espacio_fisicoTipoespacio_fisico,
+		}
+
+		c.Data["json"] = temp
+		//-------------- Temporal: Cambio por transición ------- //
+		//c.Data["json"] = v
 	}
 	c.ServeJSON()
 }
@@ -85,10 +146,10 @@ func (c *EspacioFisicoV2Controller) GetOne() {
 // @Param	order	query	string	false	"Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc ..."
 // @Param	limit	query	int	false	"Limit the size of result set. Must be an integer"
 // @Param	offset	query	int	false	"Start position of result set. Must be an integer"
-// @Success 200 {object} []models.EspacioFisicoV2
+// @Success 200 {object} []models.EspacioFisico
 // @Failure 404 not found resource
 // @router / [get]
-func (c *EspacioFisicoV2Controller) GetAll() {
+func (c *EspacioFisicoController) GetAll() {
 	var fields []string
 	var sortby []string
 	var order []string
@@ -126,7 +187,12 @@ func (c *EspacioFisicoV2Controller) GetAll() {
 				return
 			}
 			k, v := kv[0], kv[1]
-			query[k] = v
+			if k == "TipoEspacio.Id" {
+				query["TipoEspacioFisicoId.Id"] = v
+			} else {
+				query[k] = v
+			}
+
 		}
 	}
 
@@ -139,8 +205,47 @@ func (c *EspacioFisicoV2Controller) GetAll() {
 	} else {
 		if l == nil {
 			l = append(l, map[string]interface{}{})
+			c.Data["json"] = l
+		} else {
+			var temp []models.EspacioFisico
+			for _, i := range l {
+				field, _ := i.(models.EspacioFisicoV2)
+				var act string
+				if field.Activo == true {
+					act = "Activo"
+				} else {
+					act = "Inactivo"
+				}
+
+				te := &models.TipoEspacioFisico{
+					Id:                field.TipoEspacioFisicoId.Id,
+					Nombre:            field.TipoEspacioFisicoId.Nombre,
+					Descripcion:       field.TipoEspacioFisicoId.Descripcion,
+					CodigoAbreviacion: field.TipoEspacioFisicoId.CodigoAbreviacion,
+					Activo:            field.TipoEspacioFisicoId.Activo,
+					FechaCreacion:     field.TipoEspacioFisicoId.FechaCreacion,
+					FechaModificacion: field.TipoEspacioFisicoId.FechaModificacion,
+				}
+
+				x := models.EspacioFisico{
+					Id:                field.Id,
+					Nombre:            field.Nombre,
+					Codigo:            field.CodigoAbreviacion,
+					Estado:            act,
+					Descripcion:       field.Descripcion,
+					FechaCreacion:     field.FechaCreacion,
+					FechaModificacion: field.FechaModificacion,
+					TipoEspacio:       te,
+					//espacio_fisicoTipoespacio_fisico: field.espacio_fisicoTipoespacio_fisico,
+				}
+
+				temp = append(temp, x)
+			}
+			c.Data["json"] = temp
 		}
-		c.Data["json"] = l
+		//-------------- Temporal: Cambio por transición ------- //
+
+		//c.Data["json"] = l
 	}
 	c.ServeJSON()
 }
@@ -149,16 +254,34 @@ func (c *EspacioFisicoV2Controller) GetAll() {
 // @Title Put
 // @Description update the EspacioFisico
 // @Param	id		path 	int	true		"The id you want to update"
-// @Param	body		body 	models.EspacioFisicoV2	true		"body for EspacioFisico content"
-// @Success 200 {object} models.EspacioFisicoV2
+// @Param	body		body 	models.EspacioFisico	true		"body for EspacioFisico content"
+// @Success 200 {object} models.EspacioFisico
 // @Failure 400 the request contains incorrect syntax
 // @router /:id [put]
-func (c *EspacioFisicoV2Controller) Put() {
+func (c *EspacioFisicoController) Put() {
 	idStr := c.Ctx.Input.Param(":id")
 	id, _ := strconv.Atoi(idStr)
-	v := models.EspacioFisicoV2{Id: id}
+	infoDep, _ := models.GetEspacioFisicoById(id)
+	v := models.EspacioFisico{Id: id}
+
 	if err := json.Unmarshal(c.Ctx.Input.RequestBody, &v); err == nil {
-		if err := models.UpdateEspacioFisicoById(&v); err == nil {
+
+		tef := &models.TipoEspacioFisicoV2{
+			Id: v.TipoEspacio.Id,
+		}
+
+		v2 := models.EspacioFisicoV2{
+			Id:                  id,
+			Nombre:              v.Nombre,
+			Descripcion:         v.Descripcion,
+			CodigoAbreviacion:   v.Codigo,
+			Activo:              infoDep.Activo,
+			FechaCreacion:       infoDep.FechaCreacion,
+			FechaModificacion:   time.Now(),
+			TipoEspacioFisicoId: tef,
+		}
+
+		if err := models.UpdateEspacioFisicoById(&v2); err == nil {
 			c.Data["json"] = v
 		} else {
 			logs.Error(err)
@@ -182,7 +305,7 @@ func (c *EspacioFisicoV2Controller) Put() {
 // @Success 200 {object} models.Deleted
 // @Failure 404 not found resource
 // @router /:id [delete]
-func (c *EspacioFisicoV2Controller) Delete() {
+func (c *EspacioFisicoController) Delete() {
 	idStr := c.Ctx.Input.Param(":id")
 	id, _ := strconv.Atoi(idStr)
 	if err := models.DeleteEspacioFisico(id); err == nil {
@@ -199,12 +322,12 @@ func (c *EspacioFisicoV2Controller) Delete() {
 // EspaciosHuerfanos ...
 // @Title EspaciosHuerfanos
 // @Description Función para cargar los espacios físicos huerfanos
-// @Param	id		path 	int	true		"Id del espacio físico"
+// @Param	id		path 	string	true		"Id del espacio físico"
 // @Success 200 {object} []models.EspacioFisico
 // @Failure 403 id is empty
 // @router /EspaciosHuerfanos/:id [get]
 //Función para cargar los espacios físicos huerfanos
-func (c *EspacioFisicoV2Controller) EspaciosHuerfanos() {
+func (c *EspacioFisicoController) EspaciosHuerfanos() {
 	tipo := c.Ctx.Input.Param(":id")
 	id, _ := strconv.Atoi(tipo)
 	//perfiles := ("Admin_Arka")
@@ -224,7 +347,7 @@ func (c *EspacioFisicoV2Controller) EspaciosHuerfanos() {
 // @Success 200 {object} models.EspacioFisicoPadreHijo
 // @Failure 403 :espacio_fisico is empty
 // @router /get_espacios_fisicos_hijos_by_id/:espacio_fisico [get]
-func (c *EspacioFisicoV2Controller) GetEspaciosFisicosHijosById() {
+func (c *EspacioFisicoController) GetEspaciosFisicosHijosById() {
 	//Se crea variable que contiene el id con tipo de dato string
 	espacioFisicoPadre := c.Ctx.Input.Param(":espacio_fisico")
 	EFPadreint, _ := strconv.Atoi(espacioFisicoPadre)
@@ -247,11 +370,11 @@ func (c *EspacioFisicoV2Controller) GetEspaciosFisicosHijosById() {
 // GetEspaciosFisicosPadresById ...
 // @Title GetEspaciosFisicosPadresById
 // @Description A partir de una espacio_fisico dado, se obtienen todos sus predecesores en una estructura de árbol.
-// @Param	espacio_fisico	path 	string	true		"Id de la espacio_fisico"
+// @Param	espacio_fisico	path 	int	true		"Id de la espacio_fisico"
 // @Success 200 {object} []models.EspafioFisicoPadreHijo
 // @Failure 404 :espacio_fisico is empty
 // @router /get_espacios_fisicos_padres_by_id/:espacio_fisico [get]
-func (c *EspacioFisicoV2Controller) GetEspaciosFisicosPadresById() {
+func (c *EspacioFisicoController) GetEspaciosFisicosPadresById() {
 	//Se crea variable que contiene el id con tipo de dato string
 	espacioFisicoHijo := c.Ctx.Input.Param(":espacio_fisico")
 	EFHijoint, _ := strconv.Atoi(espacioFisicoHijo)
