+++ controllers/espacio_fisico_campo.go
@@ -5,20 +5,21 @@ import (
 	"errors"
 	"strconv"
 	"strings"
-
-	"github.com/udistrital/oikos_api/models"
+	"time"
 
 	"github.com/astaxie/beego"
 	"github.com/astaxie/beego/logs"
+
+	"github.com/udistrital/oikos_api/models"
 )
 
-// EspacioFisicoCampoV2Controller operations for EspacioFisicoCampo
-type EspacioFisicoCampoV2Controller struct {
+// EspacioFisicoCampoController oprations for EspacioFisicoCampo
+type EspacioFisicoCampoController struct {
 	beego.Controller
 }
 
 // URLMapping ...
-func (c *EspacioFisicoCampoV2Controller) URLMapping() {
+func (c *EspacioFisicoCampoController) URLMapping() {
 	c.Mapping("Post", c.Post)
 	c.Mapping("GetOne", c.GetOne)
 	c.Mapping("GetAll", c.GetAll)
@@ -29,14 +30,34 @@ func (c *EspacioFisicoCampoV2Controller) URLMapping() {
 // Post ...
 // @Title Post
 // @Description create EspacioFisicoCampo
-// @Param	body		body 	models.EspacioFisicoCampoV2	true		"body for EspacioFisicoCampo content"
-// @Success 201 {object} models.EspacioFisicoCampoV2
-// @Failure 400 the request contains incorrect syntax
+// @Param	body		body 	models.EspacioFisicoCampo	true		"body for EspacioFisicoCampo content"
+// @Success 201 {object} models.EspacioFisicoCampo
+// @Failure 403 body is empty
 // @router / [post]
-func (c *EspacioFisicoCampoV2Controller) Post() {
-	var v models.EspacioFisicoCampoV2
+func (c *EspacioFisicoCampoController) Post() {
+	var v models.EspacioFisicoCampo
 	if err := json.Unmarshal(c.Ctx.Input.RequestBody, &v); err == nil {
-		if _, err := models.AddEspacioFisicoCampo(&v); err == nil {
+		ca := &models.CampoV2{
+			Id: v.Campo.Id,
+		}
+
+		ef := &models.EspacioFisicoV2{
+			Id: v.EspacioFisico.Id,
+		}
+
+		temp := models.EspacioFisicoCampoV2{
+			Id:                v.Id,
+			Valor:             v.Valor,
+			CampoId:           ca,
+			EspacioFisicoId:   ef,
+			FechaInicio:       time.Now(),
+			Activo:            true,
+			FechaCreacion:     time.Now(),
+			FechaModificacion: time.Now(),
+		}
+		//-------------- Temporal: Cambio por transición ------- //
+		if _, err := models.AddEspacioFisicoCampo(&temp); err == nil {
+			//if _, err := models.AddEspacioFisicoCampo(&v); err == nil {
 			c.Ctx.Output.SetStatus(201)
 			c.Data["json"] = v
 		} else {
@@ -58,10 +79,10 @@ func (c *EspacioFisicoCampoV2Controller) Post() {
 // @Title Get One
 // @Description get EspacioFisicoCampo by id
 // @Param	id		path 	int	true		"The key for staticblock"
-// @Success 200 {object} models.EspacioFisicoCampoV2
-// @Failure 404 not found resource
+// @Success 200 {object} models.EspacioFisicoCampo
+// @Failure 403 :id is empty
 // @router /:id [get]
-func (c *EspacioFisicoCampoV2Controller) GetOne() {
+func (c *EspacioFisicoCampoController) GetOne() {
 	idStr := c.Ctx.Input.Param(":id")
 	id, _ := strconv.Atoi(idStr)
 	v, err := models.GetEspacioFisicoCampoById(id)
@@ -71,7 +92,53 @@ func (c *EspacioFisicoCampoV2Controller) GetOne() {
 		c.Data["system"] = err
 		c.Abort("404")
 	} else {
-		c.Data["json"] = v
+		//-------------- Temporal: Cambio por transición ------- //
+		te := &models.TipoEspacioFisico{
+			Id:                v.EspacioFisicoId.TipoEspacioFisicoId.Id,
+			Nombre:            v.EspacioFisicoId.TipoEspacioFisicoId.Nombre,
+			Descripcion:       v.EspacioFisicoId.TipoEspacioFisicoId.Descripcion,
+			CodigoAbreviacion: v.EspacioFisicoId.TipoEspacioFisicoId.CodigoAbreviacion,
+			Activo:            v.EspacioFisicoId.TipoEspacioFisicoId.Activo,
+			FechaCreacion:     v.EspacioFisicoId.TipoEspacioFisicoId.FechaCreacion,
+			FechaModificacion: v.EspacioFisicoId.TipoEspacioFisicoId.FechaModificacion,
+		}
+
+		ca := &models.Campo{
+			Id:                v.CampoId.Id,
+			Nombre:            v.CampoId.Nombre,
+			Descripcion:       v.CampoId.Descripcion,
+			CodigoAbreviacion: v.CampoId.CodigoAbreviacion,
+			Activo:            v.CampoId.Activo,
+			FechaCreacion:     v.CampoId.FechaCreacion,
+			FechaModificacion: v.CampoId.FechaModificacion,
+		}
+
+		ef := &models.EspacioFisico{
+			Id:                v.EspacioFisicoId.Id,
+			Nombre:            v.EspacioFisicoId.Nombre,
+			Descripcion:       v.EspacioFisicoId.Descripcion,
+			Codigo:            v.EspacioFisicoId.CodigoAbreviacion,
+			Estado:            "v.EspacioFisicoId.Activo",
+			FechaCreacion:     v.EspacioFisicoId.FechaCreacion,
+			FechaModificacion: v.EspacioFisicoId.FechaModificacion,
+			TipoEspacio:       te,
+		}
+
+		temp := models.EspacioFisicoCampo{
+			Id:                v.Id,
+			Valor:             v.Valor,
+			Campo:             ca,
+			EspacioFisico:     ef,
+			FechaInicio:       v.FechaInicio,
+			FechaFin:          v.FechaFin,
+			Activo:            v.Activo,
+			FechaCreacion:     v.FechaCreacion,
+			FechaModificacion: v.FechaModificacion,
+		}
+
+		c.Data["json"] = temp
+
+		//c.Data["json"] = v
 	}
 	c.ServeJSON()
 }
@@ -85,10 +152,10 @@ func (c *EspacioFisicoCampoV2Controller) GetOne() {
 // @Param	order	query	string	false	"Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc ..."
 // @Param	limit	query	int	false	"Limit the size of result set. Must be an integer"
 // @Param	offset	query	int	false	"Start position of result set. Must be an integer"
-// @Success 200 {object} []models.EspacioFisicoCampoV2
-// @Failure 404 not found resource
+// @Success 200 {object} []models.EspacioFisicoCampo
+// @Failure 403
 // @router / [get]
-func (c *EspacioFisicoCampoV2Controller) GetAll() {
+func (c *EspacioFisicoCampoController) GetAll() {
 	var fields []string
 	var sortby []string
 	var order []string
@@ -139,8 +206,64 @@ func (c *EspacioFisicoCampoV2Controller) GetAll() {
 	} else {
 		if l == nil {
 			l = append(l, map[string]interface{}{})
+			c.Data["json"] = l
+		} else {
+			//-------------- Temporal: Cambio por transición ------- //
+			var temp []models.EspacioFisicoCampo
+			for _, i := range l {
+				field, _ := i.(models.EspacioFisicoCampoV2)
+
+				te := &models.TipoEspacioFisico{
+					Id:                field.EspacioFisicoId.TipoEspacioFisicoId.Id,
+					Nombre:            field.EspacioFisicoId.TipoEspacioFisicoId.Nombre,
+					Descripcion:       field.EspacioFisicoId.TipoEspacioFisicoId.Descripcion,
+					CodigoAbreviacion: field.EspacioFisicoId.TipoEspacioFisicoId.CodigoAbreviacion,
+					Activo:            field.EspacioFisicoId.TipoEspacioFisicoId.Activo,
+					FechaCreacion:     field.EspacioFisicoId.TipoEspacioFisicoId.FechaCreacion,
+					FechaModificacion: field.EspacioFisicoId.TipoEspacioFisicoId.FechaModificacion,
+				}
+
+				c := &models.Campo{
+					Id:                field.CampoId.Id,
+					Nombre:            field.CampoId.Nombre,
+					Descripcion:       field.CampoId.Descripcion,
+					CodigoAbreviacion: field.CampoId.CodigoAbreviacion,
+					Activo:            field.CampoId.Activo,
+					FechaCreacion:     field.CampoId.FechaCreacion,
+					FechaModificacion: field.CampoId.FechaModificacion,
+				}
+
+				ef := &models.EspacioFisico{
+					Id:                field.EspacioFisicoId.Id,
+					Nombre:            field.EspacioFisicoId.Nombre,
+					Descripcion:       field.EspacioFisicoId.Descripcion,
+					Codigo:            field.EspacioFisicoId.CodigoAbreviacion,
+					Estado:            "field.EspacioFisicoId.Activo",
+					FechaCreacion:     field.EspacioFisicoId.FechaCreacion,
+					FechaModificacion: field.EspacioFisicoId.FechaModificacion,
+					TipoEspacio:       te,
+				}
+
+				x := models.EspacioFisicoCampo{
+					Id:                field.Id,
+					Valor:             field.Valor,
+					Campo:             c,
+					EspacioFisico:     ef,
+					FechaInicio:       field.FechaInicio,
+					FechaFin:          field.FechaFin,
+					Activo:            field.Activo,
+					FechaCreacion:     field.FechaCreacion,
+					FechaModificacion: field.FechaModificacion,
+				}
+
+				temp = append(temp, x)
+			}
+
+			c.Data["json"] = temp
+
 		}
-		c.Data["json"] = l
+
+		//c.Data["json"] = l
 	}
 	c.ServeJSON()
 }
@@ -151,12 +274,20 @@ func (c *EspacioFisicoCampoV2Controller) GetAll() {
 // @Param	id		path 	int	true		"The id you want to update"
 // @Param	body		body 	models.EspacioFisicoCampoV2	true		"body for EspacioFisicoCampo content"
 // @Success 200 {object} models.EspacioFisicoCampoV2
-// @Failure 400 the request contains incorrect syntax
+// @Failure 403 :id is not int
 // @router /:id [put]
-func (c *EspacioFisicoCampoV2Controller) Put() {
+func (c *EspacioFisicoCampoController) Put() {
 	idStr := c.Ctx.Input.Param(":id")
 	id, _ := strconv.Atoi(idStr)
-	v := models.EspacioFisicoCampoV2{Id: id}
+	//-------------- Temporal: Cambio por transición ------- //
+	infoDep, _ := models.GetEspacioFisicoCampoById(id)
+	v := models.EspacioFisicoCampoV2{
+		Id:                id,
+		Activo:            true,
+		FechaCreacion:     infoDep.FechaCreacion,
+		FechaModificacion: time.Now(),
+	}
+	//v := models.EspacioFisicoCampo{Id: id}
 	if err := json.Unmarshal(c.Ctx.Input.RequestBody, &v); err == nil {
 		if err := models.UpdateEspacioFisicoCampoById(&v); err == nil {
 			c.Data["json"] = v
@@ -179,10 +310,10 @@ func (c *EspacioFisicoCampoV2Controller) Put() {
 // @Title Delete
 // @Description delete the EspacioFisicoCampo
 // @Param	id		path 	int	true		"The id you want to delete"
-// @Success 200 {object} models.Deleted
-// @Failure 404 not found resource
+// @Success 200 {opbject} models.Deleted
+// @Failure 403 id is empty
 // @router /:id [delete]
-func (c *EspacioFisicoCampoV2Controller) Delete() {
+func (c *EspacioFisicoCampoController) Delete() {
 	idStr := c.Ctx.Input.Param(":id")
 	id, _ := strconv.Atoi(idStr)
 	if err := models.DeleteEspacioFisicoCampo(id); err == nil {
