+++ controllers/espacio_fisico_padre.go
@@ -5,20 +5,21 @@ import (
 	"errors"
 	"strconv"
 	"strings"
-
-	"github.com/udistrital/oikos_api/models"
+	"time"
 
 	"github.com/astaxie/beego"
 	"github.com/astaxie/beego/logs"
+
+	"github.com/udistrital/oikos_api/models"
 )
 
-// EspacioFisicoPadreV2Controller operations for EspacioFisicoPadre
-type EspacioFisicoPadreV2Controller struct {
+// EspacioFisicoPadreController oprations for EspacioFisicoPadre
+type EspacioFisicoPadreController struct {
 	beego.Controller
 }
 
 // URLMapping ...
-func (c *EspacioFisicoPadreV2Controller) URLMapping() {
+func (c *EspacioFisicoPadreController) URLMapping() {
 	c.Mapping("Post", c.Post)
 	c.Mapping("GetOne", c.GetOne)
 	c.Mapping("GetAll", c.GetAll)
@@ -29,14 +30,33 @@ func (c *EspacioFisicoPadreV2Controller) URLMapping() {
 // Post ...
 // @Title Post
 // @Description create EspacioFisicoPadre
-// @Param	body		body 	models.EspacioFisicoPadreV2	true		"body for EspacioFisicoPadre content"
-// @Success 201 {object} models.EspacioFisicoPadreV2
+// @Param	body		body 	models.EspacioFisicoPadre	true		"body for EspacioFisicoPadre content"
+// @Success 201 {object} models.EspacioFisicoPadre
 // @Failure 400 the request contains incorrect syntax
 // @router / [post]
-func (c *EspacioFisicoPadreV2Controller) Post() {
-	var v models.EspacioFisicoPadreV2
+func (c *EspacioFisicoPadreController) Post() {
+	var v models.EspacioFisicoPadre
 	if err := json.Unmarshal(c.Ctx.Input.RequestBody, &v); err == nil {
-		if _, err := models.AddEspacioFisicoPadre(&v); err == nil {
+		//-------------- Temporal: Cambio por transición ------- //
+
+		efp := &models.EspacioFisicoV2{
+			Id: v.Padre.Id,
+		}
+
+		efh := &models.EspacioFisicoV2{
+			Id: v.Hijo.Id,
+		}
+
+		temp := models.EspacioFisicoPadreV2{
+			Id:                v.Id,
+			PadreId:           efp,
+			HijoId:            efh,
+			FechaCreacion:     time.Now(),
+			FechaModificacion: time.Now(),
+		}
+		//-------------- Temporal: Cambio por transición ------- //
+		if _, err := models.AddEspacioFisicoPadre(&temp); err == nil {
+			//if _, err := models.AddEspacioFisicoPadre(&v); err == nil {
 			c.Ctx.Output.SetStatus(201)
 			c.Data["json"] = v
 		} else {
@@ -58,10 +78,10 @@ func (c *EspacioFisicoPadreV2Controller) Post() {
 // @Title Get One
 // @Description get EspacioFisicoPadre by id
 // @Param	id		path 	int	true		"The key for staticblock"
-// @Success 200 {object} models.EspacioFisicoPadreV2
+// @Success 200 {object} models.EspacioFisicoPadre
 // @Failure 404 not found resource
 // @router /:id [get]
-func (c *EspacioFisicoPadreV2Controller) GetOne() {
+func (c *EspacioFisicoPadreController) GetOne() {
 	idStr := c.Ctx.Input.Param(":id")
 	id, _ := strconv.Atoi(idStr)
 	v, err := models.GetEspacioFisicoPadreById(id)
@@ -71,7 +91,26 @@ func (c *EspacioFisicoPadreV2Controller) GetOne() {
 		c.Data["system"] = err
 		c.Abort("404")
 	} else {
-		c.Data["json"] = v
+		//-------------- Temporal: Cambio por transición ------- //
+
+		efp := &models.EspacioFisico{
+			Id: v.PadreId.Id,
+		}
+
+		efh := &models.EspacioFisico{
+			Id: v.HijoId.Id,
+		}
+
+		temp := models.EspacioFisicoPadre{
+			Id:                v.Id,
+			Padre:             efp,
+			Hijo:              efh,
+			FechaCreacion:     v.FechaCreacion,
+			FechaModificacion: v.FechaModificacion,
+		}
+
+		c.Data["json"] = temp
+		//c.Data["json"] = v
 	}
 	c.ServeJSON()
 }
@@ -83,12 +122,12 @@ func (c *EspacioFisicoPadreV2Controller) GetOne() {
 // @Param	fields	query	string	false	"Fields returned. e.g. col1,col2 ..."
 // @Param	sortby	query	string	false	"Sorted-by fields. e.g. col1,col2 ..."
 // @Param	order	query	string	false	"Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc ..."
-// @Param	limit	query	string	int	"Limit the size of result set. Must be an integer"
-// @Param	offset	query	string	int	"Start position of result set. Must be an integer"
-// @Success 200 {object} []models.EspacioFisicoPadreV2
+// @Param	limit	query	int	false	"Limit the size of result set. Must be an integer"
+// @Param	offset	query	int	false	"Start position of result set. Must be an integer"
+// @Success 200 {object}  []models.EspacioFisicoPadre
 // @Failure 404 not found resource
 // @router / [get]
-func (c *EspacioFisicoPadreV2Controller) GetAll() {
+func (c *EspacioFisicoPadreController) GetAll() {
 	var fields []string
 	var sortby []string
 	var order []string
@@ -139,8 +178,82 @@ func (c *EspacioFisicoPadreV2Controller) GetAll() {
 	} else {
 		if l == nil {
 			l = append(l, map[string]interface{}{})
+			c.Data["json"] = l
+		} else {
+			//-------------- Temporal: Cambio por transición ------- //
+			var temp []models.EspacioFisicoPadre
+			var act string
+			for _, i := range l {
+				field, _ := i.(models.EspacioFisicoPadreV2)
+
+				tefp := &models.TipoEspacioFisico{
+					Id:                field.PadreId.TipoEspacioFisicoId.Id,
+					Nombre:            field.PadreId.TipoEspacioFisicoId.Nombre,
+					Descripcion:       field.PadreId.TipoEspacioFisicoId.Descripcion,
+					CodigoAbreviacion: field.PadreId.TipoEspacioFisicoId.CodigoAbreviacion,
+					Activo:            field.PadreId.TipoEspacioFisicoId.Activo,
+					FechaCreacion:     field.PadreId.TipoEspacioFisicoId.FechaCreacion,
+					FechaModificacion: field.PadreId.TipoEspacioFisicoId.FechaModificacion,
+				}
+
+				tefh := &models.TipoEspacioFisico{
+					Id:                field.HijoId.TipoEspacioFisicoId.Id,
+					Nombre:            field.HijoId.TipoEspacioFisicoId.Nombre,
+					Descripcion:       field.HijoId.TipoEspacioFisicoId.Descripcion,
+					CodigoAbreviacion: field.HijoId.TipoEspacioFisicoId.CodigoAbreviacion,
+					Activo:            field.HijoId.TipoEspacioFisicoId.Activo,
+					FechaCreacion:     field.HijoId.TipoEspacioFisicoId.FechaCreacion,
+					FechaModificacion: field.HijoId.TipoEspacioFisicoId.FechaModificacion,
+				}
+
+				if field.PadreId.Activo == true {
+					act = "Activo"
+				} else {
+					act = "Inactivo"
+				}
+
+				efp := &models.EspacioFisico{
+					Id:                field.PadreId.Id,
+					Nombre:            field.PadreId.Nombre,
+					Estado:            act,
+					Codigo:            field.PadreId.CodigoAbreviacion,
+					Descripcion:       field.PadreId.Descripcion,
+					FechaCreacion:     field.PadreId.FechaCreacion,
+					FechaModificacion: field.PadreId.FechaModificacion,
+					TipoEspacio:       tefp,
+				}
+
+				if field.HijoId.Activo == true {
+					act = "Activo"
+				} else {
+					act = "Inactivo"
+				}
+
+				efh := &models.EspacioFisico{
+					Id:                field.HijoId.Id,
+					Nombre:            field.HijoId.Nombre,
+					Estado:            act,
+					Codigo:            field.HijoId.CodigoAbreviacion,
+					Descripcion:       field.HijoId.Descripcion,
+					FechaCreacion:     field.HijoId.FechaCreacion,
+					FechaModificacion: field.HijoId.FechaModificacion,
+					TipoEspacio:       tefh,
+				}
+
+				x := models.EspacioFisicoPadre{
+					Id:                field.Id,
+					Padre:             efp,
+					Hijo:              efh,
+					FechaCreacion:     field.FechaCreacion,
+					FechaModificacion: field.FechaModificacion,
+				}
+
+				temp = append(temp, x)
+			}
+			c.Data["json"] = temp
 		}
-		c.Data["json"] = l
+
+		//c.Data["json"] = l
 	}
 	c.ServeJSON()
 }
@@ -149,16 +262,33 @@ func (c *EspacioFisicoPadreV2Controller) GetAll() {
 // @Title Put
 // @Description update the EspacioFisicoPadre
 // @Param	id		path 	int	true		"The id you want to update"
-// @Param	body		body 	models.EspacioFisicoPadreV2	true		"body for EspacioFisicoPadre content"
-// @Success 200 {object} models.EspacioFisicoPadreV2
+// @Param	body		body 	models.EspacioFisicoPadre	true		"body for EspacioFisicoPadre content"
+// @Success 200 {object} models.EspacioFisicoPadre
 // @Failure 400 the request contains incorrect syntax
 // @router /:id [put]
-func (c *EspacioFisicoPadreV2Controller) Put() {
+func (c *EspacioFisicoPadreController) Put() {
 	idStr := c.Ctx.Input.Param(":id")
 	id, _ := strconv.Atoi(idStr)
-	v := models.EspacioFisicoPadreV2{Id: id}
+	v := models.EspacioFisicoPadre{Id: id}
+
 	if err := json.Unmarshal(c.Ctx.Input.RequestBody, &v); err == nil {
-		if err := models.UpdateEspacioFisicoPadreById(&v); err == nil {
+		//-------------- Temporal: Cambio por transición ------- //
+		efp := &models.EspacioFisicoV2{
+			Id: v.Padre.Id,
+		}
+		efh := &models.EspacioFisicoV2{
+			Id: v.Hijo.Id,
+		}
+
+		v2 := models.EspacioFisicoPadreV2{
+			Id:                id,
+			PadreId:           efp,
+			HijoId:            efh,
+			FechaCreacion:     v.FechaCreacion,
+			FechaModificacion: time.Now(),
+		}
+
+		if err := models.UpdateEspacioFisicoPadreById(&v2); err == nil {
 			c.Data["json"] = v
 		} else {
 			logs.Error(err)
@@ -182,7 +312,7 @@ func (c *EspacioFisicoPadreV2Controller) Put() {
 // @Success 200 {object} models.Deleted
 // @Failure 404 not found resource
 // @router /:id [delete]
-func (c *EspacioFisicoPadreV2Controller) Delete() {
+func (c *EspacioFisicoPadreController) Delete() {
 	idStr := c.Ctx.Input.Param(":id")
 	id, _ := strconv.Atoi(idStr)
 	if err := models.DeleteEspacioFisicoPadre(id); err == nil {
