+++ controllers/espacio_fisico_v2.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 	"strconv"
 	"strings"
-	"time"
 
 	"github.com/astaxie/beego"
 	"github.com/astaxie/beego/logs"
@@ -13,13 +12,13 @@ import (
 	"github.com/udistrital/oikos_api/models"
 )
 
-// DependenciaTipoDependenciaController oprations for DependenciaTipoDependencia
-type DependenciaTipoDependenciaController struct {
+// EspacioFisicoV2Controller operations for EspacioFisico
+type EspacioFisicoV2Controller struct {
 	beego.Controller
 }
 
 // URLMapping ...
-func (c *DependenciaTipoDependenciaController) URLMapping() {
+func (c *EspacioFisicoV2Controller) URLMapping() {
 	c.Mapping("Post", c.Post)
 	c.Mapping("GetOne", c.GetOne)
 	c.Mapping("GetAll", c.GetAll)
@@ -29,26 +28,16 @@ func (c *DependenciaTipoDependenciaController) URLMapping() {
 
 // Post ...
 // @Title Post
-// @Description create DependenciaTipoDependencia
-// @Param	body		body 	models.DependenciaTipoDependencia	true		"body for DependenciaTipoDependencia content"
-// @Success 201 {object} models.DependenciaTipoDependencia
+// @Description create EspacioFisico
+// @Param	body		body 	models.EspacioFisicoV2	true		"body for EspacioFisico content"
+// @Success 201 {object} models.EspacioFisicoV2
 // @Failure 400 the request contains incorrect syntax
 // @router / [post]
-func (c *DependenciaTipoDependenciaController) Post() {
-	var v models.DependenciaTipoDependencia
+func (c *EspacioFisicoV2Controller) Post() {
+	var v models.EspacioFisicoV2
 	if err := json.Unmarshal(c.Ctx.Input.RequestBody, &v); err == nil {
-		//-------------- Temporal: Cambio por transición ------- //
-		var temp models.DependenciaTipoDependenciaV2
-		temp.FromV1(v)
-		temp.Activo = true
-		t := time.Now()
-		temp.FechaCreacion = t
-		temp.FechaModificacion = t
-		//-------------- Temporal: Cambio por transición ------- //
-		if _, err := models.AddDependenciaTipoDependencia(&temp); err == nil {
-			//if _, err := models.AddDependenciaTipoDependencia(&v); err == nil {
+		if _, err := models.AddEspacioFisico(&v); err == nil {
 			c.Ctx.Output.SetStatus(201)
-			temp.ToV1(&v)
 			c.Data["json"] = v
 		} else {
 			logs.Error(err)
@@ -67,44 +56,39 @@ func (c *DependenciaTipoDependenciaController) Post() {
 
 // GetOne ...
 // @Title Get One
-// @Description get DependenciaTipoDependencia by id
+// @Description get EspacioFisico by id
 // @Param	id		path 	int	true		"The key for staticblock"
-// @Success 200 {object} models.DependenciaTipoDependencia
+// @Success 200 {object} models.EspacioFisicoV2
 // @Failure 404 not found resource
 // @router /:id [get]
-func (c *DependenciaTipoDependenciaController) GetOne() {
+func (c *EspacioFisicoV2Controller) GetOne() {
 	idStr := c.Ctx.Input.Param(":id")
 	id, _ := strconv.Atoi(idStr)
-	v, err := models.GetDependenciaTipoDependenciaById(id)
+	v, err := models.GetEspacioFisicoById(id)
 	if err != nil {
 		logs.Error(err)
 		//c.Data["development"] = map[string]interface{}{"Code": "000", "Body": err.Error(), "Type": "error"}
 		c.Data["system"] = err
 		c.Abort("404")
 	} else {
-		//-------------- Temporal: Cambio por transición ------- //
-		var temp models.DependenciaTipoDependencia
-		v.ToV1(&temp)
-		c.Data["json"] = temp
-
-		//c.Data["json"] = v
+		c.Data["json"] = v
 	}
 	c.ServeJSON()
 }
 
 // GetAll ...
 // @Title Get All
-// @Description get DependenciaTipoDependencia
+// @Description get EspacioFisico
 // @Param	query	query	string	false	"Filter. e.g. col1:v1,col2:v2 ..."
 // @Param	fields	query	string	false	"Fields returned. e.g. col1,col2 ..."
 // @Param	sortby	query	string	false	"Sorted-by fields. e.g. col1,col2 ..."
 // @Param	order	query	string	false	"Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc ..."
 // @Param	limit	query	int	false	"Limit the size of result set. Must be an integer"
 // @Param	offset	query	int	false	"Start position of result set. Must be an integer"
-// @Success 200 {object} []models.DependenciaTipoDependencia
+// @Success 200 {object} []models.EspacioFisicoV2
 // @Failure 404 not found resource
 // @router / [get]
-func (c *DependenciaTipoDependenciaController) GetAll() {
+func (c *EspacioFisicoV2Controller) GetAll() {
 	var fields []string
 	var sortby []string
 	var order []string
@@ -146,7 +130,7 @@ func (c *DependenciaTipoDependenciaController) GetAll() {
 		}
 	}
 
-	l, err := models.GetAllDependenciaTipoDependencia(query, fields, sortby, order, offset, limit)
+	l, err := models.GetAllEspacioFisico(query, fields, sortby, order, offset, limit)
 	if err != nil {
 		logs.Error(err)
 		//c.Data["development"] = map[string]interface{}{"Code": "000", "Body": err.Error(), "Type": "error"}
@@ -155,52 +139,26 @@ func (c *DependenciaTipoDependenciaController) GetAll() {
 	} else {
 		if l == nil {
 			l = append(l, map[string]interface{}{})
-			c.Data["json"] = l
-		} else {
-			//-------------- Temporal: Cambio por transición ------- //
-			var temp []interface{}
-			for _, i := range l {
-				switch v := i.(type) {
-				case map[string]interface{}:
-					temp = append(temp, v)
-				case models.DependenciaTipoDependenciaV2:
-					var x models.DependenciaTipoDependencia
-					v.ToV1(&x)
-					temp = append(temp, x)
-					// default:
-					// 	// SIN MANEJAR!
-				}
-			}
-			c.Data["json"] = temp
 		}
-
-		//c.Data["json"] = l
+		c.Data["json"] = l
 	}
 	c.ServeJSON()
 }
 
 // Put ...
 // @Title Put
-// @Description update the DependenciaTipoDependencia
+// @Description update the EspacioFisico
 // @Param	id		path 	int	true		"The id you want to update"
-// @Param	body		body 	models.DependenciaTipoDependenciaV2	true		"body for DependenciaTipoDependencia content"
-// @Success 200 {object} models.DependenciaTipoDependenciaV2
+// @Param	body		body 	models.EspacioFisicoV2	true		"body for EspacioFisico content"
+// @Success 200 {object} models.EspacioFisicoV2
 // @Failure 400 the request contains incorrect syntax
 // @router /:id [put]
-func (c *DependenciaTipoDependenciaController) Put() {
+func (c *EspacioFisicoV2Controller) Put() {
 	idStr := c.Ctx.Input.Param(":id")
 	id, _ := strconv.Atoi(idStr)
-	v2, _ := models.GetDependenciaTipoDependenciaById(id)
-	v := models.DependenciaTipoDependencia{Id: id}
-	//-------------- Temporal: Cambio por transición ------- //
-	// TODO: Revisar lo siguiente ...:
-	v2.Id = id
-	v2.FechaModificacion = time.Now()
-	// ... debería bastar con:
-	// v2.FromV1(v)
+	v := models.EspacioFisicoV2{Id: id}
 	if err := json.Unmarshal(c.Ctx.Input.RequestBody, &v); err == nil {
-		if err := models.UpdateDependenciaTipoDependenciaById(v2); err == nil {
-			v2.ToV1(&v)
+		if err := models.UpdateEspacioFisicoById(&v); err == nil {
 			c.Data["json"] = v
 		} else {
 			logs.Error(err)
@@ -219,21 +177,96 @@ func (c *DependenciaTipoDependenciaController) Put() {
 
 // Delete ...
 // @Title Delete
-// @Description delete the DependenciaTipoDependencia
+// @Description delete the EspacioFisico
 // @Param	id		path 	int	true		"The id you want to delete"
 // @Success 200 {object} models.Deleted
 // @Failure 404 not found resource
 // @router /:id [delete]
-func (c *DependenciaTipoDependenciaController) Delete() {
+func (c *EspacioFisicoV2Controller) Delete() {
 	idStr := c.Ctx.Input.Param(":id")
 	id, _ := strconv.Atoi(idStr)
-	if err := models.DeleteDependenciaTipoDependencia(id); err == nil {
-		c.Data["json"] = models.Deleted{Id: id}
+	if err := models.DeleteEspacioFisico(id); err == nil {
+		c.Data["json"] = map[string]interface{}{"Id": id}
 	} else {
 		logs.Error(err)
 		//c.Data["development"] = map[string]interface{}{"Code": "000", "Body": err.Error(), "Type": "error"}
 		c.Data["system"] = err
 		c.Abort("404")
 	}
+	c.ServeJSON()
+}
+
+// EspaciosHuerfanos ...
+// @Title EspaciosHuerfanos
+// @Description Función para cargar los espacios físicos huerfanos
+// @Param	id		path 	int	true		"Id del espacio físico"
+// @Success 200 {object} []models.EspacioFisico
+// @Failure 403 id is empty
+// @router /EspaciosHuerfanos/:id [get]
+//Función para cargar los espacios físicos huerfanos
+func (c *EspacioFisicoV2Controller) EspaciosHuerfanos() {
+	tipo := c.Ctx.Input.Param(":id")
+	id, _ := strconv.Atoi(tipo)
+	//perfiles := ("Admin_Arka")
+	//perfilesR := strings.NewReplacer(",", "','")
+
+	//Construcción Json Menús Huerfanos
+	l := models.EspacioFisicosHuerfanos(id)
+	c.Data["json"] = l
+	//Generera el Json con los datos obtenidos
+	c.ServeJSON()
+}
+
+// GetEspaciosFisicosHijosById ...
+// @Title GetEspaciosFisicosHijosById
+// @Description A partir de un espacio físico dado, se obtienen los hijas de él en una estructura de árbol.
+// @Param	espacio_fisico	path 	int	true		"Id del espacio físico"
+// @Success 200 {object} models.EspacioFisicoPadreHijo
+// @Failure 403 :espacio_fisico is empty
+// @router /get_espacios_fisicos_hijos_by_id/:espacio_fisico [get]
+func (c *EspacioFisicoV2Controller) GetEspaciosFisicosHijosById() {
+	//Se crea variable que contiene el id con tipo de dato string
+	espacioFisicoPadre := c.Ctx.Input.Param(":espacio_fisico")
+	EFPadreint, _ := strconv.Atoi(espacioFisicoPadre)
+	l, err := models.GetEspaciosFisicosHijosById(EFPadreint)
+	if err != nil {
+		beego.Error(err)
+		//c.Data["development"] = map[string]interface{}{"Code": "000", "Body": err.Error(), "Type": "error"}
+		c.Data["system"] = err
+		c.Abort("404")
+
+	} else {
+
+		c.Data["json"] = map[string]interface{}{"Body": l, "Type": "success"}
+	}
+
+	//Generera el Json con los datos obtenidos
+	c.ServeJSON()
+}
+
+// GetEspaciosFisicosPadresById ...
+// @Title GetEspaciosFisicosPadresById
+// @Description A partir de una espacio_fisico dado, se obtienen todos sus predecesores en una estructura de árbol.
+// @Param	espacio_fisico	path 	string	true		"Id de la espacio_fisico"
+// @Success 200 {object} []models.EspafioFisicoPadreHijo
+// @Failure 404 :espacio_fisico is empty
+// @router /get_espacios_fisicos_padres_by_id/:espacio_fisico [get]
+func (c *EspacioFisicoV2Controller) GetEspaciosFisicosPadresById() {
+	//Se crea variable que contiene el id con tipo de dato string
+	espacioFisicoHijo := c.Ctx.Input.Param(":espacio_fisico")
+	EFHijoint, _ := strconv.Atoi(espacioFisicoHijo)
+	l, err := models.GetEspaciosFisicosPadresById(EFHijoint)
+	if err != nil {
+		beego.Error(err)
+		//c.Data["development"] = map[string]interface{}{"Code": "000", "Body": err.Error(), "Type": "error"}
+		c.Data["system"] = err
+		c.Abort("404")
+
+	} else {
+
+		c.Data["json"] = map[string]interface{}{"Body": l, "Type": "success"}
+	}
+
+	//Generera el Json con los datos obtenidos
 	c.ServeJSON()
 }
