+++ controllers/tipo_dependencia.go
@@ -5,20 +5,21 @@ import (
 	"errors"
 	"strconv"
 	"strings"
-
-	"github.com/udistrital/oikos_api/models"
+	"time"
 
 	"github.com/astaxie/beego"
 	"github.com/astaxie/beego/logs"
+
+	"github.com/udistrital/oikos_api/models"
 )
 
-// TipoDependenciaV2Controller operations for TipoDependencia
-type TipoDependenciaV2Controller struct {
+// TipoDependenciaController oprations for TipoDependencia
+type TipoDependenciaController struct {
 	beego.Controller
 }
 
 // URLMapping ...
-func (c *TipoDependenciaV2Controller) URLMapping() {
+func (c *TipoDependenciaController) URLMapping() {
 	c.Mapping("Post", c.Post)
 	c.Mapping("GetOne", c.GetOne)
 	c.Mapping("GetAll", c.GetAll)
@@ -29,14 +30,26 @@ func (c *TipoDependenciaV2Controller) URLMapping() {
 // Post ...
 // @Title Post
 // @Description create TipoDependencia
-// @Param	body		body 	models.TipoDependenciaV2	true		"body for TipoDependencia content"
-// @Success 201 {object} models.TipoDependenciaV2
+// @Param	body		body 	models.TipoDependencia	true		"body for TipoDependencia content"
+// @Success 201 {object} models.TipoDependencia
 // @Failure 400 the request contains incorrect syntax
 // @router / [post]
-func (c *TipoDependenciaV2Controller) Post() {
-	var v models.TipoDependenciaV2
+func (c *TipoDependenciaController) Post() {
+	var v models.TipoDependencia
 	if err := json.Unmarshal(c.Ctx.Input.RequestBody, &v); err == nil {
-		if _, err := models.AddTipoDependencia(&v); err == nil {
+		//-------------- Temporal: Cambio por transición ------- //
+		temp := models.TipoDependenciaV2{
+			Id:                v.Id,
+			Nombre:            v.Nombre,
+			Descripcion:       "Descripción",
+			CodigoAbreviacion: "TU_" + v.Nombre,
+			Activo:            true,
+			FechaCreacion:     time.Now(),
+			FechaModificacion: time.Now(),
+		}
+		//-------------- Temporal: Cambio por transición ------- //
+		if _, err := models.AddTipoDependencia(&temp); err == nil {
+			//if _, err := models.AddTipoDependencia(&v); err == nil {
 			c.Ctx.Output.SetStatus(201)
 			c.Data["json"] = v
 		} else {
@@ -58,10 +71,10 @@ func (c *TipoDependenciaV2Controller) Post() {
 // @Title Get One
 // @Description get TipoDependencia by id
 // @Param	id		path 	int	true		"The key for staticblock"
-// @Success 200 {object} models.TipoDependenciaV2
+// @Success 200 {object} models.TipoDependencia
 // @Failure 404 not found resource
 // @router /:id [get]
-func (c *TipoDependenciaV2Controller) GetOne() {
+func (c *TipoDependenciaController) GetOne() {
 	idStr := c.Ctx.Input.Param(":id")
 	id, _ := strconv.Atoi(idStr)
 	v, err := models.GetTipoDependenciaById(id)
@@ -71,7 +84,20 @@ func (c *TipoDependenciaV2Controller) GetOne() {
 		c.Data["system"] = err
 		c.Abort("404")
 	} else {
-		c.Data["json"] = v
+		//-------------- Temporal: Cambio por transición ------- //
+
+		temp := models.TipoDependencia{
+			Id:                v.Id,
+			Nombre:            v.Nombre,
+			Descripcion:       v.Descripcion,
+			CodigoAbreviacion: v.CodigoAbreviacion,
+			Activo:            v.Activo,
+			FechaCreacion:     v.FechaCreacion,
+			FechaModificacion: v.FechaModificacion,
+		}
+		c.Data["json"] = temp
+		//-------------- Temporal: Cambio por transición ------- //
+		//c.Data["json"] = v
 	}
 	c.ServeJSON()
 }
@@ -85,10 +111,10 @@ func (c *TipoDependenciaV2Controller) GetOne() {
 // @Param	order	query	string	false	"Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc ..."
 // @Param	limit	query	int	false	"Limit the size of result set. Must be an integer"
 // @Param	offset	query	int	false	"Start position of result set. Must be an integer"
-// @Success 200 {object} []models.TipoDependenciaV2
+// @Success 200 {object} []models.TipoDependencia
 // @Failure 404 not found resource
 // @router / [get]
-func (c *TipoDependenciaV2Controller) GetAll() {
+func (c *TipoDependenciaController) GetAll() {
 	var fields []string
 	var sortby []string
 	var order []string
@@ -139,8 +165,29 @@ func (c *TipoDependenciaV2Controller) GetAll() {
 	} else {
 		if l == nil {
 			l = append(l, map[string]interface{}{})
+			c.Data["json"] = l
+		} else {
+			//-------------- Temporal: Cambio por transición ------- //
+			var temp []models.TipoDependencia
+			for _, i := range l {
+				field, _ := i.(models.TipoDependenciaV2)
+				x := models.TipoDependencia{
+					Id:                field.Id,
+					Nombre:            field.Nombre,
+					Descripcion:       field.Descripcion,
+					CodigoAbreviacion: field.CodigoAbreviacion,
+					Activo:            field.Activo,
+					FechaCreacion:     field.FechaCreacion,
+					FechaModificacion: field.FechaModificacion,
+				}
+
+				temp = append(temp, x)
+			}
+			c.Data["json"] = temp
 		}
-		c.Data["json"] = l
+
+		//-------------- Temporal: Cambio por transición ------- //
+		//c.Data["json"] = l
 	}
 	c.ServeJSON()
 }
@@ -149,16 +196,27 @@ func (c *TipoDependenciaV2Controller) GetAll() {
 // @Title Put
 // @Description update the TipoDependencia
 // @Param	id		path 	int	true		"The id you want to update"
-// @Param	body		body 	models.TipoDependenciaV2	true		"body for TipoDependencia content"
-// @Success 200 {object} models.TipoDependenciaV2
+// @Param	body		body 	models.TipoDependencia	true		"body for TipoDependencia content"
+// @Success 200 {object} models.TipoDependencia
 // @Failure 400 the request contains incorrect syntax
 // @router /:id [put]
-func (c *TipoDependenciaV2Controller) Put() {
+func (c *TipoDependenciaController) Put() {
 	idStr := c.Ctx.Input.Param(":id")
 	id, _ := strconv.Atoi(idStr)
-	v := models.TipoDependenciaV2{Id: id}
+	//-------------- Temporal: Cambio por transición ------- //
+	infoDep, _ := models.GetTipoDependenciaById(id)
+	v := models.TipoDependencia{Id: id}
 	if err := json.Unmarshal(c.Ctx.Input.RequestBody, &v); err == nil {
-		if err := models.UpdateTipoDependenciaById(&v); err == nil {
+		v2 := models.TipoDependenciaV2{
+			Id:                id,
+			Nombre:            v.Nombre,
+			Descripcion:       infoDep.Descripcion,
+			CodigoAbreviacion: infoDep.CodigoAbreviacion,
+			Activo:            infoDep.Activo,
+			FechaCreacion:     infoDep.FechaCreacion,
+			FechaModificacion: time.Now(),
+		}
+		if err := models.UpdateTipoDependenciaById(&v2); err == nil {
 			c.Data["json"] = v
 		} else {
 			logs.Error(err)
@@ -182,7 +240,7 @@ func (c *TipoDependenciaV2Controller) Put() {
 // @Success 200 {object} models.Deleted
 // @Failure 404 not found resource
 // @router /:id [delete]
-func (c *TipoDependenciaV2Controller) Delete() {
+func (c *TipoDependenciaController) Delete() {
 	idStr := c.Ctx.Input.Param(":id")
 	id, _ := strconv.Atoi(idStr)
 	if err := models.DeleteTipoDependencia(id); err == nil {
