+++ controllers/tipo_dependencia_v2.go
@@ -5,21 +5,20 @@ import (
 	"errors"
 	"strconv"
 	"strings"
-	"time"
+
+	"github.com/udistrital/oikos_api/models"
 
 	"github.com/astaxie/beego"
 	"github.com/astaxie/beego/logs"
-
-	"github.com/udistrital/oikos_api/models"
 )
 
-// TipoDependenciaController oprations for TipoDependencia
-type TipoDependenciaController struct {
+// TipoDependenciaV2Controller operations for TipoDependencia
+type TipoDependenciaV2Controller struct {
 	beego.Controller
 }
 
 // URLMapping ...
-func (c *TipoDependenciaController) URLMapping() {
+func (c *TipoDependenciaV2Controller) URLMapping() {
 	c.Mapping("Post", c.Post)
 	c.Mapping("GetOne", c.GetOne)
 	c.Mapping("GetAll", c.GetAll)
@@ -30,33 +29,15 @@ func (c *TipoDependenciaController) URLMapping() {
 // Post ...
 // @Title Post
 // @Description create TipoDependencia
-// @Param	body		body 	models.TipoDependencia	true		"body for TipoDependencia content"
-// @Success 201 {object} models.TipoDependencia
+// @Param	body		body 	models.TipoDependenciaV2	true		"body for TipoDependencia content"
+// @Success 201 {object} models.TipoDependenciaV2
 // @Failure 400 the request contains incorrect syntax
 // @router / [post]
-func (c *TipoDependenciaController) Post() {
-	var v models.TipoDependencia
+func (c *TipoDependenciaV2Controller) Post() {
+	var v models.TipoDependenciaV2
 	if err := json.Unmarshal(c.Ctx.Input.RequestBody, &v); err == nil {
-		//-------------- Temporal: Cambio por transición ------- //
-
-		// TODO: Revisar lo siguiente ...:
-		temp := models.TipoDependenciaV2{
-			Id:                v.Id,
-			Nombre:            v.Nombre,
-			Descripcion:       "Descripción",
-			CodigoAbreviacion: "TU_" + v.Nombre,
-			Activo:            true,
-			FechaCreacion:     time.Now(),
-			FechaModificacion: time.Now(),
-		}
-		// ... debería bastar con:
-		// var temp models.TipoDependenciaV2
-		// temp.FromV1(v)
-		//-------------- Temporal: Cambio por transición ------- //
-		if _, err := models.AddTipoDependencia(&temp); err == nil {
-			//if _, err := models.AddTipoDependencia(&v); err == nil {
+		if _, err := models.AddTipoDependencia(&v); err == nil {
 			c.Ctx.Output.SetStatus(201)
-			temp.ToV1(&v)
 			c.Data["json"] = v
 		} else {
 			logs.Error(err)
@@ -77,10 +58,10 @@ func (c *TipoDependenciaController) Post() {
 // @Title Get One
 // @Description get TipoDependencia by id
 // @Param	id		path 	int	true		"The key for staticblock"
-// @Success 200 {object} models.TipoDependencia
+// @Success 200 {object} models.TipoDependenciaV2
 // @Failure 404 not found resource
 // @router /:id [get]
-func (c *TipoDependenciaController) GetOne() {
+func (c *TipoDependenciaV2Controller) GetOne() {
 	idStr := c.Ctx.Input.Param(":id")
 	id, _ := strconv.Atoi(idStr)
 	v, err := models.GetTipoDependenciaById(id)
@@ -90,10 +71,7 @@ func (c *TipoDependenciaController) GetOne() {
 		c.Data["system"] = err
 		c.Abort("404")
 	} else {
-		//-------------- Temporal: Cambio por transición ------- //
-		var temp models.TipoDependencia
-		v.ToV1(&temp)
-		c.Data["json"] = temp
+		c.Data["json"] = v
 	}
 	c.ServeJSON()
 }
@@ -107,10 +85,10 @@ func (c *TipoDependenciaController) GetOne() {
 // @Param	order	query	string	false	"Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc ..."
 // @Param	limit	query	int	false	"Limit the size of result set. Must be an integer"
 // @Param	offset	query	int	false	"Start position of result set. Must be an integer"
-// @Success 200 {object} []models.TipoDependencia
+// @Success 200 {object} []models.TipoDependenciaV2
 // @Failure 404 not found resource
 // @router / [get]
-func (c *TipoDependenciaController) GetAll() {
+func (c *TipoDependenciaV2Controller) GetAll() {
 	var fields []string
 	var sortby []string
 	var order []string
@@ -161,24 +139,8 @@ func (c *TipoDependenciaController) GetAll() {
 	} else {
 		if l == nil {
 			l = append(l, map[string]interface{}{})
-			c.Data["json"] = l
-		} else {
-			//-------------- Temporal: Cambio por transición ------- //
-			var temp []interface{}
-			for _, i := range l {
-				switch v := i.(type) {
-				case map[string]interface{}:
-					temp = append(temp, v)
-				case models.TipoDependenciaV2:
-					var x models.TipoDependencia
-					v.ToV1(&x)
-					temp = append(temp, v)
-					// default:
-					// 	// SIN MANEJAR!
-				}
-			}
-			c.Data["json"] = temp
 		}
+		c.Data["json"] = l
 	}
 	c.ServeJSON()
 }
@@ -187,25 +149,16 @@ func (c *TipoDependenciaController) GetAll() {
 // @Title Put
 // @Description update the TipoDependencia
 // @Param	id		path 	int	true		"The id you want to update"
-// @Param	body		body 	models.TipoDependencia	true		"body for TipoDependencia content"
-// @Success 200 {object} models.TipoDependencia
+// @Param	body		body 	models.TipoDependenciaV2	true		"body for TipoDependencia content"
+// @Success 200 {object} models.TipoDependenciaV2
 // @Failure 400 the request contains incorrect syntax
 // @router /:id [put]
-func (c *TipoDependenciaController) Put() {
+func (c *TipoDependenciaV2Controller) Put() {
 	idStr := c.Ctx.Input.Param(":id")
 	id, _ := strconv.Atoi(idStr)
-	//-------------- Temporal: Cambio por transición ------- //
-	v2, _ := models.GetTipoDependenciaById(id)
-	v := models.TipoDependencia{Id: id}
+	v := models.TipoDependenciaV2{Id: id}
 	if err := json.Unmarshal(c.Ctx.Input.RequestBody, &v); err == nil {
-		// TODO: Revisar lo siguiente...:
-		v2.Id = id
-		v2.Nombre = v.Nombre
-		v2.FechaModificacion = time.Now()
-		// ... debería bastar con:
-		// v2.FromV1(v)
-		if err := models.UpdateTipoDependenciaById(v2); err == nil {
-			v2.ToV1(&v)
+		if err := models.UpdateTipoDependenciaById(&v); err == nil {
 			c.Data["json"] = v
 		} else {
 			logs.Error(err)
@@ -229,11 +182,11 @@ func (c *TipoDependenciaController) Put() {
 // @Success 200 {object} models.Deleted
 // @Failure 404 not found resource
 // @router /:id [delete]
-func (c *TipoDependenciaController) Delete() {
+func (c *TipoDependenciaV2Controller) Delete() {
 	idStr := c.Ctx.Input.Param(":id")
 	id, _ := strconv.Atoi(idStr)
 	if err := models.DeleteTipoDependencia(id); err == nil {
-		c.Data["json"] = models.Deleted{Id: id}
+		c.Data["json"] = map[string]interface{}{"Id": id}
 	} else {
 		logs.Error(err)
 		//c.Data["development"] = map[string]interface{}{"Code": "000", "Body": err.Error(), "Type": "error"}
