+++ controllers/tipo_espacio_fisico_v2.go
@@ -5,21 +5,20 @@ import (
 	"errors"
 	"strconv"
 	"strings"
-	"time"
+
+	"github.com/udistrital/oikos_api/models"
 
 	"github.com/astaxie/beego"
 	"github.com/astaxie/beego/logs"
-
-	"github.com/udistrital/oikos_api/models"
 )
 
-// TipoEspacioFisicoController oprations for TipoEspacioFisico
-type TipoEspacioFisicoController struct {
+// TipoEspacioFisicoV2Controller operations for TipoEspacioFisico
+type TipoEspacioFisicoV2Controller struct {
 	beego.Controller
 }
 
 // URLMapping ...
-func (c *TipoEspacioFisicoController) URLMapping() {
+func (c *TipoEspacioFisicoV2Controller) URLMapping() {
 	c.Mapping("Post", c.Post)
 	c.Mapping("GetOne", c.GetOne)
 	c.Mapping("GetAll", c.GetAll)
@@ -30,27 +29,14 @@ func (c *TipoEspacioFisicoController) URLMapping() {
 // Post ...
 // @Title Post
 // @Description create TipoEspacioFisico
-// @Param	body		body 	models.TipoEspacioFisico	true		"body for TipoEspacioFisico content"
-// @Success 201 {object} models.TipoEspacioFisico
+// @Param	body		body 	models.TipoEspacioFisicoV2	true		"body for TipoEspacioFisico content"
+// @Success 201 {object} models.TipoEspacioFisicoV2
 // @Failure 400 the request contains incorrect syntax
 // @router / [post]
-func (c *TipoEspacioFisicoController) Post() {
-	var v models.TipoEspacioFisico
+func (c *TipoEspacioFisicoV2Controller) Post() {
+	var v models.TipoEspacioFisicoV2
 	if err := json.Unmarshal(c.Ctx.Input.RequestBody, &v); err == nil {
-		//-------------- Temporal: Cambio por transición ------- //
-
-		temp := models.TipoEspacioFisicoV2{
-			Id:                v.Id,
-			Nombre:            v.Nombre,
-			Descripcion:       "Descripción",
-			CodigoAbreviacion: "TU_" + v.Nombre,
-			Activo:            true,
-			FechaCreacion:     time.Now(),
-			FechaModificacion: time.Now(),
-		}
-		//-------------- Temporal: Cambio por transición ------- //
-		if _, err := models.AddTipoEspacioFisico(&temp); err == nil {
-			//if _, err := models.AddTipoEspacioFisico(&v); err == nil {
+		if _, err := models.AddTipoEspacioFisico(&v); err == nil {
 			c.Ctx.Output.SetStatus(201)
 			c.Data["json"] = v
 		} else {
@@ -72,10 +58,10 @@ func (c *TipoEspacioFisicoController) Post() {
 // @Title Get One
 // @Description get TipoEspacioFisico by id
 // @Param	id		path 	int	true		"The key for staticblock"
-// @Success 200 {object} models.TipoEspacioFisico
+// @Success 200 {object} models.TipoEspacioFisicoV2
 // @Failure 404 not found resource
 // @router /:id [get]
-func (c *TipoEspacioFisicoController) GetOne() {
+func (c *TipoEspacioFisicoV2Controller) GetOne() {
 	idStr := c.Ctx.Input.Param(":id")
 	id, _ := strconv.Atoi(idStr)
 	v, err := models.GetTipoEspacioFisicoById(id)
@@ -85,20 +71,7 @@ func (c *TipoEspacioFisicoController) GetOne() {
 		c.Data["system"] = err
 		c.Abort("404")
 	} else {
-		//-------------- Temporal: Cambio por transición ------- //
-
-		temp := models.TipoEspacioFisico{
-			Id:                v.Id,
-			Nombre:            v.Nombre,
-			Descripcion:       v.Descripcion,
-			CodigoAbreviacion: v.CodigoAbreviacion,
-			Activo:            v.Activo,
-			FechaCreacion:     v.FechaCreacion,
-			FechaModificacion: v.FechaModificacion,
-		}
-		c.Data["json"] = temp
-		//-------------- Temporal: Cambio por transición ------- //
-		//c.Data["json"] = v
+		c.Data["json"] = v
 	}
 	c.ServeJSON()
 }
@@ -112,10 +85,10 @@ func (c *TipoEspacioFisicoController) GetOne() {
 // @Param	order	query	string	false	"Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc ..."
 // @Param	limit	query	int	false	"Limit the size of result set. Must be an integer"
 // @Param	offset	query	int	false	"Start position of result set. Must be an integer"
-// @Success 200 {object} []models.TipoEspacioFisico
+// @Success 200 {object} []models.TipoEspacioFisicoV2
 // @Failure 404 not found resource
 // @router / [get]
-func (c *TipoEspacioFisicoController) GetAll() {
+func (c *TipoEspacioFisicoV2Controller) GetAll() {
 	var fields []string
 	var sortby []string
 	var order []string
@@ -166,31 +139,8 @@ func (c *TipoEspacioFisicoController) GetAll() {
 	} else {
 		if l == nil {
 			l = append(l, map[string]interface{}{})
-			c.Data["json"] = l
-		} else {
-			//-------------- Temporal: Cambio por transición ------- //
-			var temp []models.TipoEspacioFisico
-			for _, i := range l {
-				field, _ := i.(models.TipoEspacioFisicoV2)
-				x := models.TipoEspacioFisico{
-					Id:                field.Id,
-					Nombre:            field.Nombre,
-					Descripcion:       field.Descripcion,
-					CodigoAbreviacion: field.CodigoAbreviacion,
-					Activo:            field.Activo,
-					FechaCreacion:     field.FechaCreacion,
-					FechaModificacion: field.FechaModificacion,
-				}
-
-				temp = append(temp, x)
-			}
-
-			c.Data["json"] = temp
 		}
-
-		//-------------- Temporal: Cambio por transición ------- //
-
-		//c.Data["json"] = l
+		c.Data["json"] = l
 	}
 	c.ServeJSON()
 }
@@ -199,28 +149,16 @@ func (c *TipoEspacioFisicoController) GetAll() {
 // @Title Put
 // @Description update the TipoEspacioFisico
 // @Param	id		path 	int	true		"The id you want to update"
-// @Param	body		body 	models.TipoEspacioFisico	true		"body for TipoEspacioFisico content"
-// @Success 200 {object} models.TipoEspacioFisico
+// @Param	body		body 	models.TipoEspacioFisicoV2	true		"body for TipoEspacioFisico content"
+// @Success 200 {object} models.TipoEspacioFisicoV2
 // @Failure 400 the request contains incorrect syntax
 // @router /:id [put]
-func (c *TipoEspacioFisicoController) Put() {
+func (c *TipoEspacioFisicoV2Controller) Put() {
 	idStr := c.Ctx.Input.Param(":id")
 	id, _ := strconv.Atoi(idStr)
-	//-------------- Temporal: Cambio por transición ------- //
-	infoDep, _ := models.GetTipoEspacioFisicoById(id)
-	v := models.TipoEspacioFisico{Id: id}
+	v := models.TipoEspacioFisicoV2{Id: id}
 	if err := json.Unmarshal(c.Ctx.Input.RequestBody, &v); err == nil {
-		v2 := models.TipoEspacioFisicoV2{
-			Id:                id,
-			Nombre:            v.Nombre,
-			Descripcion:       infoDep.Descripcion,
-			CodigoAbreviacion: infoDep.CodigoAbreviacion,
-			Activo:            infoDep.Activo,
-			FechaCreacion:     infoDep.FechaCreacion,
-			FechaModificacion: time.Now(),
-		}
-
-		if err := models.UpdateTipoEspacioFisicoById(&v2); err == nil {
+		if err := models.UpdateTipoEspacioFisicoById(&v); err == nil {
 			c.Data["json"] = v
 		} else {
 			logs.Error(err)
@@ -244,7 +182,7 @@ func (c *TipoEspacioFisicoController) Put() {
 // @Success 200 {object} models.Deleted
 // @Failure 404 not found resource
 // @router /:id [delete]
-func (c *TipoEspacioFisicoController) Delete() {
+func (c *TipoEspacioFisicoV2Controller) Delete() {
 	idStr := c.Ctx.Input.Param(":id")
 	id, _ := strconv.Atoi(idStr)
 	if err := models.DeleteTipoEspacioFisico(id); err == nil {
